% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spomic.R
\docType{class}
\name{Spomic-class}
\alias{Spomic-class}
\title{Spomic Class}
\description{
The \code{Spomic} S4 class is designed to store details, data, preprocessing steps,
and results for spatial analysis of single-cell spatial omics.
This structure makes it easy to keep track of hyperparameters,
input data, transformations, and output in one object.
}
\section{Slots}{

\describe{
\item{\code{details}}{A \code{list} containing metadata or hyperparameters of various functions.}

\item{\code{df}}{A \code{data.frame} containing the main dataset.}

\item{\code{pp}}{An object of any type (\code{spp}) containing preprocessing steps.}

\item{\code{results}}{A \code{list} containing analysis results, such as fitted models, metrics, or plots.}
}}

